chore: renaming recorder redux to TMDefinition
