--- src/redux/Recorder/selectors.ts
@@ -1,3 +1,0 @@
-import { IRootState } from "@redux/store";
-
-export const recorderSelector = (state: IRootState) => state.recorder;
