+++ src/redux/TMDefinition/selectors.ts
@@ -1,0 +1,3 @@
+import { IRootState } from "@redux/store";
+
+export const TMDefinitionSelector = (state: IRootState) => state.TMDefinition;
